backend/requirements.txt
anyio==4.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
Faker==25.9.1
h11==0.16.0
httptools==0.6.4
idna==3.10
mariadb==1.1.13
langchain-ollama
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.4.0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
faiss-cpu>=1.7.4
numpy>=1.24.0
langgraph
langchain-community
langchain-core
Shapely
